snowpipe: extract out schema evolution to seperate struct
This will help to re-use all this logic when we create the new output
that specifies channel names explicitly.
